requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sqlite4
pydantic==2.5.0
python-multipart==0.0.6
playwright==1.55.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
aiosqlite==0.21.0
redis==6.4.0

# AI/ML libraries
ultralytics==8.2.0
anthropic==0.7.7
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.4
torch==2.4.1
torchvision==0.19.1

